% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbinaryGMM.R
\name{getSummary.bingmm}
\alias{getSummary.bingmm}
\title{Get Model Summaries for use with "mtable" for objects of class bingmm}
\usage{
\method{getSummary}{bingmm}(
  obj,
  alpha = 0.05,
  vce = c("robust", "efficient", "ml"),
  method = "bhhh",
  R = 1000,
  ...
)
}
\arguments{
\item{obj}{a \code{bingmm} object,}

\item{alpha}{level of the confidence intervals,}

\item{...}{further arguments,}
}
\description{
A generic function to collect coefficients and summary statistics from a \code{bingmm} object. It is used in \code{mtable}
}
\details{
For more details see package \pkg{memisc}.
}
