% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbinaryRis.R
\name{getSummary.binris}
\alias{getSummary.binris}
\title{Get Model Summaries for use with "mtable" for objects of class binris}
\usage{
\method{getSummary}{binris}(obj, alpha = 0.05, ...)
}
\arguments{
\item{obj}{a \code{binris} object,}

\item{alpha}{level of the confidence intervals,}

\item{...}{further arguments,}
}
\value{
A list with an array with coefficient estimates and a vector containing the model summary statistics.
}
\description{
A generic function to collect coefficients and summary statistics from a \code{binris} object. It is used in \code{mtable}
}
\details{
For more details see package \pkg{memisc}.
}
