% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbinaryRis.R
\name{getSummary.maxLik}
\alias{getSummary.maxLik}
\title{Get Model Summaries for use with "mtable" for objects of class maxLik}
\usage{
\method{getSummary}{maxLik}(obj, alpha = 0.05, ...)
}
\arguments{
\item{obj}{a \code{maxLik} object,}

\item{alpha}{level of the confidence intervals,}

\item{...}{further arguments,}
}
\description{
A generic function to collect coefficients and summary statistics from a \code{maxLik} object. It is used in \code{mtable}
}
\details{
For more details see package \pkg{memisc}.
}
