% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalEffects.R
\name{impacts}
\alias{impacts}
\title{Estimation of the average marginal effects for SARB models.}
\usage{
impacts(object, ...)
}
\arguments{
\item{object}{an object of class \code{bingmm} or \code{binlgmm}}

\item{...}{Additional arguments to be passed.}
}
\value{
Estimates of the direct, indirect and total effect.
}
\description{
Obtain the average marginal effects from \code{bingmm} or \code{binlgmm} class model.
}
