% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbinaryGMM.R
\name{make.instruments}
\alias{make.instruments}
\title{Make instruments for spatial models}
\usage{
make.instruments(listw, x, q)
}
\arguments{
\item{listw}{object. An object of class \code{listw}, \code{matrix}, or \code{Matrix}.}

\item{x}{variable(s) to be lagged}

\item{q}{number of lags}
}
\description{
Make instruments for spatial models
}
\author{
Mauricio Sarrias and Gianfranco Piras.
}
\keyword{instruments}
