% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalEffects.R
\name{impacts.bingmm}
\alias{impacts.bingmm}
\alias{impacts}
\alias{impacts.binlgmm}
\alias{impacts.binris}
\alias{print.impacts.bingmm}
\alias{summary.impacts.bingmm}
\alias{print.summary.impacts.bingmm}
\title{Compute Marginal Effects for Spatial Binary  Models}
\usage{
\method{impacts}{bingmm}(
  obj,
  data = NULL,
  variable = NULL,
  from.unit = 1,
  dydx = c("exact", "numeric"),
  change = NULL,
  vce = c("robust", "efficient", "ml"),
  type = c("delta", "mc"),
  result = c("summary", "from.region", "cumulative"),
  vcov = NULL,
  het = TRUE,
  empirical = FALSE,
  approximation = FALSE,
  R = 50,
  pw = 5,
  draws = NULL,
  Q = 3,
  tol = 1e-06,
  verbose = FALSE,
  ...
)

\method{impacts}{binlgmm}(
  obj,
  data = NULL,
  variable = NULL,
  from.unit = 1,
  dydx = c("exact", "numeric"),
  change = NULL,
  vce = c("robust", "efficient", "ml"),
  type = c("delta", "mc"),
  result = c("summary", "from.region", "cumulative"),
  vcov = NULL,
  het = TRUE,
  empirical = FALSE,
  approximation = FALSE,
  R = 50,
  pw = 5,
  draws = NULL,
  Q = 3,
  tol = 1e-06,
  verbose = FALSE,
  ...
)

\method{impacts}{binris}(
  obj,
  data = NULL,
  variable = NULL,
  from.unit = 1,
  dydx = c("exact", "numeric"),
  change = NULL,
  vce = c("robust", "efficient", "ml"),
  type = c("delta", "mc"),
  result = c("summary", "from.region", "cumulative"),
  vcov = NULL,
  het = TRUE,
  empirical = FALSE,
  approximation = FALSE,
  R = 50,
  pw = 5,
  draws = NULL,
  Q = 3,
  tol = 1e-06,
  verbose = FALSE,
  ...
)

\method{print}{impacts.bingmm}(x, ...)

\method{summary}{impacts.bingmm}(object, ...)

\method{print}{summary.impacts.bingmm}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{obj}{An object of class \code{bingmm}, \code{binlgmm} or \code{binris}.}

\item{data}{Optional dataset used for computing partial effects.}

\item{variable}{Variable name (string) for which marginal effects are computed.}

\item{from.unit}{Integer specifying the spatial unit (row) for computing impacts.}

\item{dydx}{Character: type of derivative approximation to use (\code{"exact"} or \code{"numeric"}).}

\item{change}{For numeric variables, a vector indicating the lower and upper value to compute the partial change.}

\item{vce}{A string indicating what kind of variance-covariance matrix of the estimate should be computed when using \code{effect.bingmm}. For the one-step GMM estimator, the options are \code{"robust"} and \code{"ml"}. For the two-step GMM estimator, the options are \code{"robust"}, \code{"efficient"} and \code{"ml"}. The option \code{"vce = ml"} is an exploratory method that evaluates the VC of the RIS estimator using the GMM estimates.}

\item{type}{String indicating which method is used to compute the standard errors of the marginal effects. If \code{"mc"}, then the Monte Carlo approximation is used. If \code{"delta"}, then the Delta Method is used.}

\item{result}{String indicating the output format: \code{"summary"} presents the average direct, indirect or total effects; \code{"from.region"} display the total effect of the variable in \code{"variable"} form spatial unit indicated in \code{"from.unit"}; if \code{"cumulative"} the cumulative effect up to power order indicated in \code{"Q"} for the variable in \code{"variable"} form spatial unit indicated in \code{"from.unit"} is returned.}

\item{vcov}{Optional user-supplied variance-covariance matrix.}

\item{het}{Logical. If \code{TRUE} (the default), then the heteroskedasticity is taken into account when computing the marginal effects.}

\item{empirical}{Logical. Argument passed to \code{mvrnorm} (default \code{FALSE}): if \code{TRUE}, the coefficients and their covariance matrix specify the empirical not population mean and covariance matrix.}

\item{approximation}{Logical. If \code{TRUE} then \eqn{(I - \lambda W)^{-1}} is approximated as \eqn{I + \lambda W + \lambda^2 W^2 + \lambda^3 W^3 + ... +\lambda^q W^q}. The default is \code{FALSE}.}

\item{R}{Numerical. Indicates the number of draws used in the Monte Carlo approximation if \code{type = "mc"}.}

\item{pw}{numeric. The power used for the approximation \eqn{I + \lambda W + \lambda^2 W^2 + \lambda^3 W^3 + ... +\lambda^q W^q}. The default is 5.}

\item{draws}{Optional matrix of parameter draws for Monte Carlo simulation.}

\item{Q}{Number of columns for cumulative effects (if applicable).}

\item{tol}{Numerical. Argument passed to \code{mvrnorm}: tolerance (relative to largest variance) for numerical lack of positive-definiteness in the coefficient covariance matrix.}

\item{verbose}{Logical. Display progress.}

\item{...}{further arguments. Ignored.}

\item{x}{An object of class \code{impacts.bingmm} for \code{print} methods.}

\item{object}{An object of class \code{impacts.bingmm} for \code{summary} methods.}

\item{digits}{the number of digits.}
}
\value{
An object of class \code{impacts.bingmm}.
}
\description{
Computes marginal effects (direct, indirect, total, from-region, and cumulative) from a spatial binary response model.
Supports inference via the Delta method or Monte Carlo simulation.
}
\details{
Let the model be:

\deqn{
y^*= X\beta + WX\gamma + \lambda W y^* + \epsilon = Z\delta + \lambda Wy^{*} + \epsilon
}

where  \eqn{y = 1} if \eqn{y^*>0} and 0 otherwise; \eqn{\epsilon \sim N(0, 1)} if \code{link = "probit"} or \eqn{\epsilon \sim L(0, \pi^2/3)} if \code{link = "logit"}. 
The RIS estimator assumes that \eqn{\epsilon \sim N(0, 1)}. 

The marginal effects respect to variable \eqn{x_r} can be computed as

\deqn{
diag(f(a))D^{-1}_{\lambda}A^{-1}_{\lambda}\left(I_n\beta_r + W\gamma_r\right) = C_r(\theta)
}

where \eqn{f()} is the pdf, which depends on the assumption of the error terms; \eqn{diag} is the operator that creates a \eqn{n \times n} diagonal matrix; \eqn{A_{\lambda}= (I -\lambda W)}; and \eqn{D_{\lambda}} is a diagonal matrix whose elements represent the square root of the diagonal elements of the variance-covariance matrix of \eqn{u = A_{\lambda}^{-1}\epsilon}. 

We implement these three summary measures: (1) The average total effects, \eqn{ATE_r  = n^{-1}i_n'C_{r}i_n}, (2) The average direct effects, \eqn{ADE_r  = n^{-1}tr(C_{r})}, and (3) the average indirect effects, \eqn{ATE_r - ADE_r}. 

The standard errors of the average total, direct and indirect effects can be estimated using either Monte Carlo (MC) approximation, which takes into account the sampling distribution of \eqn{\theta}, or Delta Method.
}
\examples{
\donttest{
# Data set
data(oldcol, package = "spdep")

# Create dependent (dummy) variable
COL.OLD$CRIMED <- as.numeric(COL.OLD$CRIME > 35)

# Two-step (Probit) GMM estimator
ts <- sbinaryGMM(CRIMED ~ INC * HOVAL + factor(CP) | HOVAL,
                link = "probit", 
                listw = spdep::nb2listw(COL.nb, style = "W"), 
                data = COL.OLD, 
                type = "twostep")
                
# Summary marginal effects using Delta Method
summary(impacts(ts, type = "delta"))

# Summary marginal effects using MC with 100 draws
summary(impacts(ts, type = "mc", R = 100))

# Marginal effects using efficient VC matrix
summary(impacts(ts, type = "delta", vce = "efficient"))

# Marginal effects using efficient VC matrix and ignoring the heteroskedasticity
summary(impacts(ts, type = "delta", vce = "efficient", het = FALSE))

# Obtain total impact when INC changes in region 20
imp <- impacts(ts, type = "delta", result = "from.region", from.unit = 20, variable = "INC")
head(imp$out, 6)

# Cumulative impact when INC changes in region 20 up to power = 3
cum <- impacts(ts, type = "delta", result = "cumulative", from.unit = 20, variable = "INC", Q = 3)
head(cum$out$`Q:3`, 6)

# Marginal effects using  RIS estimator
ris_sar <- sbinaryRis(CRIMED ~ INC + HOVAL, data = COL.OLD,
                      R = 50, 
                      listw = spdep::nb2listw(COL.nb, style = "W"))
summary(impacts(ris_sar, method = "delta"))
summary(impacts(ris_sar, method = "mc", R = 100))
}
}
\seealso{
\code{\link[spldv]{sbinaryGMM}}, \code{\link[spldv]{sbinaryLGMM}}.
}
\author{
Mauricio Sarrias and Gianfranco Piras.
}
\keyword{effects}
\keyword{marginal}
