% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbinaryGMM.R
\name{sbinaryGMM}
\alias{sbinaryGMM}
\alias{coef.bingmm}
\alias{vcov.bingmm}
\alias{print.bingmm}
\alias{summary.bingmm}
\alias{print.summary.bingmm}
\title{Estimation of SAR for binary models using GMM}
\usage{
sbinaryGMM(
  formula,
  data,
  subset,
  na.action,
  listw = NULL,
  nins = 2,
  link = c("logit", "probit"),
  winitial = c("optimal", "identity"),
  s.matrix = c("iid", "robust"),
  type = c("onestep", "twostep"),
  gradient = TRUE,
  start = NULL,
  cons.opt = FALSE,
  approximation = FALSE,
  verbose = TRUE,
  print.init = FALSE,
  ...
)

\method{coef}{bingmm}(object, ...)

\method{vcov}{bingmm}(
  object,
  vce = c("robust", "efficient", "ml"),
  method = "bhhh",
  R = 1000,
  ...
)

\method{print}{bingmm}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{bingmm}(
  object,
  vce = c("robust", "efficient", "ml"),
  method = "bhhh",
  R = 1000,
  ...
)

\method{print}{summary.bingmm}(x, digits = max(5, getOption("digits") - 3), ...)
}
\description{
Estimation of SAR for binary models using GMM
}
