% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbinaryLGMM.R
\name{sbinaryLGMM}
\alias{sbinaryLGMM}
\alias{vcov.binlgmm}
\alias{print.binlgmm}
\alias{summary.binlgmm}
\alias{print.summary.binlgmm}
\title{Estimation of SAR for binary models using LGMM}
\usage{
sbinaryLGMM(
  formula,
  data,
  subset,
  na.action,
  listw = NULL,
  nins = 2,
  link = c("logit", "probit"),
  ...
)

\method{vcov}{binlgmm}(object, ...)

\method{print}{binlgmm}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{binlgmm}(object, ...)

\method{print}{summary.binlgmm}(
  x,
  digits = max(3, getOption("digits") - 2),
  width = getOption("width"),
  ...
)
}
\description{
Estimation of SAR for binary models using LGMM
}
