% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbinaryRis.R
\name{sbinaryRis}
\alias{sbinaryRis}
\title{Estimation of SAR for binary models using RIS}
\usage{
sbinaryRis(
  formula,
  data,
  subset,
  na.action,
  listw = NULL,
  R = 1000,
  method = "bhhh",
  rho.init = NULL,
  print.init = TRUE,
  ll = c("ll_ris", "ll_ris2"),
  eigen = FALSE,
  ...
)
}
\description{
Estimation of SAR for binary models using RIS
}
